sctp: make outq layer control when to flush/xmit the packet
With this change, outq layer can know that the packet is full and
needs to be sent, and that a packet with data was transmitted. With that
now it can select a transport before queueing the next chunk in the
current packet, or select the next stream to dequeue from.

This change is an enabler for both SCTP_SS_RR_PKT and CMT-SCTP.