sctp: replace SFR-CACC by SFR
SFR-CACC[1] was an algorithm proposed for preventing unnecessary fast
retransmissions due to reordering during path failover.

It works quite well if you consider sending data only over a single
transport at a time, but it doesn't if you consider CMT-SCTP.  For
CMT-SCTP the reordering is inherent and a new algorithm is necessary,
one that is not tied to path failovers.

This patch thus replaces SFR-CACC with SFR[2]. SFR requires the tracking
of on which transport the chunk was sent (which we already do) and based
on that it can correctly determine if the Gap seen (possibly received on
another path) really means that a given TSN seems lost or not by
checking the highest TSN acked *for the path that the chunk was sent*.

This algorithm can be considered as a generalization of SFR-CACC and
thus that original issue (reordering during failover) is also covered by
it.

Note that SFR does not apply to retransmitted chunks. As retransmissions
may happen on other transports, it is impossible to do this transport
matching for all situations. One possible improvement, however, is to
detect unnecessary fast retransmissions by checking if the ACK arrived
before the estimated path delay on the retransmitted path. (Which I
think we already do)

[1] https://tools.ietf.org/html/draft-iyengar-sctp-cacc-03
[2] https://tools.ietf.org/html/draft-tuexen-tsvwg-sctp-multipath-15#section-3.1